Use a single impl block per type
